backend_fix/requirements.txt
fastapi==0.122.0
uvicorn==0.22.0
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3.0
pydantic-settings>=2.12.0
python-dotenv>=1.0.0
google-generativeai==0.8.5